requirements.txt
asgiref==3.7.2
Django==3.2.19
django-cors-headers==4.1.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-gis==1.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
python-dotenv==0.21.1
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
